BACKPORT: tcp: fix tcp_rearm_rto()
skbs in (re)transmit queue no longer have a copy of jiffies
at the time of the transmit : skb->skb_mstamp is now in usec unit,
with no correlation to tcp_jiffies32.

We have to convert rto from jiffies to usec, compute a time difference
in usec, then convert the delta to HZ units.

saalim :- Adapt to latest upstream
* from https://github.com/danascape/kernel-msm-4.14/commit/b17b8a20c5cd4a264601eacf1fda29008047d05a

Fixes: bac041c1fc8822e ("tcp: switch TCP TS option (RFC 7323) to 1ms clock")
Signed-off-by: Eric Dumazet <edumazet@google.com>
Acked-by: Soheil Hassas Yeganeh <soheil@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: danascape <saalimquadri1@gmail.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>
Signed-off-by: Zymboee <80381223+Zymboee@users.noreply.github.com>